requirement.txt
certifi==2019.11.28
chardet==3.0.4
Django==2.2.7
django-cors-headers==3.2.1
django-jsonfield==1.4.0
djangorestframework==3.11.0
idna==2.8
Pillow==6.2.1
psycopg2-binary==2.8.4
pytz==2019.3
requests==2.22.0
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.7
